# examples/demo-openapi.yaml
openapi: 3.0.3
info:
  title: Product-to-MCP Demo Store
  version: 1.0.0
paths:
  /products:
    get:
      operationId: listProducts
      summary: List products
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Product list
    post:
      operationId: createProduct
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name, price]
              properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
      responses:
        '201':
          description: Created
  /products/{product_id}:
    get:
      operationId: getProduct
      summary: Get one product
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product
    put:
      operationId: replaceProduct
      summary: Replace product
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name, price]
              properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
      responses:
        '200':
          description: Replaced
    patch:
      operationId: updateProduct
      summary: Update product
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
      responses:
        '200':
          description: Updated
    delete:
      operationId: deleteProduct
      summary: Delete product
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted